Add unit tests, refactor
